+++ packages/drizzle/src/factory/input.ts
@@ -18,15 +18,20 @@ import {
   GraphQLString,
   isNonNullType,
 } from "graphql"
+import { isColumnVisible } from "../helper"
 import { DrizzleWeaver } from "../index"
+import type { DrizzleResolverFactoryOptions } from "../types"
 
 export class DrizzleInputFactory<TTable extends Table> {
-  public constructor(public readonly table: TTable) {}
+  public constructor(
+    public readonly table: TTable,
+    public readonly options?: DrizzleResolverFactoryOptions<TTable>
+  ) {}
 
   public selectArrayArgs() {
     const name = `${pascalCase(getTableName(this.table))}SelectArrayArgs`
     const existing = weaverContext.getNamedType(name) as GraphQLObjectType
-    if (existing) return existing
+    if (existing != null) return existing
 
     return weaverContext.memoNamedType(
       new GraphQLObjectType<SelectArrayArgs<TTable>>({
@@ -136,7 +141,12 @@ export class DrizzleInputFactory<TTable extends Table> {
     return weaverContext.memoNamedType(
       new GraphQLObjectType({
         name,
-        fields: mapValue(columns, (column) => {
+        fields: mapValue(columns, (column, columnName) => {
+          if (
+            !isColumnVisible(columnName, this.options?.input ?? {}, "insert")
+          ) {
+            return mapValue.SKIP
+          }
           const type = (() => {
             const t = DrizzleWeaver.getColumnType(column)
             if (column.hasDefault) return t
@@ -160,7 +170,12 @@ export class DrizzleInputFactory<TTable extends Table> {
     return weaverContext.memoNamedType(
       new GraphQLObjectType({
         name,
-        fields: mapValue(columns, (column) => {
+        fields: mapValue(columns, (column, columnName) => {
+          if (
+            !isColumnVisible(columnName, this.options?.input ?? {}, "update")
+          ) {
+            return mapValue.SKIP
+          }
           const type = DrizzleWeaver.getColumnType(column)
           return { type }
         }),
@@ -177,9 +192,17 @@ export class DrizzleInputFactory<TTable extends Table> {
     const filterFields: Record<
       string,
       GraphQLFieldConfig<any, any, any>
-    > = mapValue(columns, (column) => ({
-      type: DrizzleInputFactory.columnFilters(column),
-    }))
+    > = mapValue(columns, (column, columnName) => {
+      if (
+        isColumnVisible(columnName, this.options?.input ?? {}, "filters") ===
+        false
+      ) {
+        return mapValue.SKIP
+      }
+      return {
+        type: DrizzleInputFactory.columnFilters(column),
+      }
+    })
 
     const filtersOr = new GraphQLObjectType({
       name: `${pascalCase(getTableName(this.table))}FiltersOr`,