+++ packages/drizzle/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## next (YYYY-MM-DD)
 
+- Feat: Add column visibility configuration for input and filter generation
+
 ## 0.8.2 (2025-04-08)
 
 - Feat: support drizzle text enum